Remove Tabindex Attribute
The tabindex attribute on the search widget and comment section
shouldn't be added by the parent theme, there are a lot of arguments not
to use it at all since the browser will naturally tab through the
document structure.

On top of that, the parent theme doesn't know how the child theme will
handle the layout and shouldn't be handling it for that reason also.